refactored modules from dir `framework`: moved to other dirs, removed
