+++ access_log/test_access_log_h2.py
@@ -6,8 +6,8 @@ __license__ = "GPL2"
 
 import os
 
-from framework.x509 import CertGenerator
 from helpers import dmesg, remote, tf_cfg
+from helpers.x509 import CertGenerator
 
 from .common import AccessLogLine
