+++ cache/test_conditional.py
@@ -6,8 +6,8 @@ __license__ = "GPL2"
 
 from framework.deproxy_client import DeproxyClient
 from framework.deproxy_server import StaticDeproxyServer
-from framework.parameterize import param, parameterize, parameterize_class
 from helpers.deproxy import HttpMessage
+from test_suite.parameterize import param, parameterize, parameterize_class
 from test_suite.tester import TempestaTest
 
 DEPROXY_CLIENT = {
