+++ framework/__init__.py
@@ -4,8 +4,6 @@ __all__ = [
     "deproxy_manager",
     "deproxy_server",
     "nginx_server",
-    "templates",
-    "tester",
     "wrk_client",
     "port_checks",
     "external_client",
