+++ http2_general/test_h2_headers.py
@@ -7,14 +7,13 @@ analises its return code.
 import itertools
 
 from h2.errors import ErrorCodes
-from h2.stream import StreamInputs
 from hyperframe import frame
 
-from framework.parameterize import param, parameterize, parameterize_class
 from helpers import tf_cfg
 from helpers.deproxy import HttpMessage
 from http2_general.helpers import H2Base
 from test_suite import tester
+from test_suite.parameterize import param, parameterize, parameterize_class
 
 __author__ = "Tempesta Technologies, Inc."
 __copyright__ = "Copyright (C) 2023-2024 Tempesta Technologies, Inc."
