+++ t_fault_injection/test_fault_injection_base.py
@@ -4,12 +4,10 @@ __author__ = "Tempesta Technologies, Inc."
 __copyright__ = "Copyright (C) 2024 Tempesta Technologies, Inc."
 __license__ = "GPL2"
 
-from framework.parameterize import param, parameterize
-from framework.x509 import CertGenerator
 from helpers import deproxy, dmesg, remote, sysnet, tf_cfg
-from helpers.deproxy import HttpMessage
 from helpers.networker import NetWorker
 from test_suite import tester
+from test_suite.parameterize import param, parameterize
 
 
 class TestFailFunctionBase(tester.TempestaTest):
