+++ t_frang/test_config.py
@@ -4,11 +4,8 @@ __author__ = "Tempesta Technologies, Inc."
 __copyright__ = "Copyright (C) 2024 Tempesta Technologies, Inc."
 __license__ = "GPL2"
 
-import time
-
 from hyperframe.frame import DataFrame, HeadersFrame
 
-from framework.parameterize import param, parameterize
 from helpers.dmesg import (
     amount_positive,
     limited_rate_on_tempesta_node,
@@ -16,6 +13,7 @@ from helpers.dmesg import (
 )
 from helpers.remote import CmdError
 from test_suite import tester
+from test_suite.parameterize import param, parameterize
 
 
 class TestDefaultConfig(tester.TempestaTest):
