+++ t_frang/test_http_resp_code_block.py
@@ -15,8 +15,8 @@ __author__ = "Tempesta Technologies, Inc."
 __copyright__ = "Copyright (C) 2022-2023 Tempesta Technologies, Inc."
 __license__ = "GPL2"
 
-from framework.parameterize import parameterize_class
 from t_frang.frang_test_case import FrangTestCase, H2Config
+from test_suite.parameterize import parameterize_class
 
 NGINX_CONFIG = {
     "id": "nginx",
