+++ t_frang/test_length.py
@@ -4,8 +4,8 @@ __author__ = "Tempesta Technologies, Inc."
 __copyright__ = "Copyright (C) 2022-2023 Tempesta Technologies, Inc."
 __license__ = "GPL2"
 
-from framework.parameterize import param, parameterize
 from t_frang.frang_test_case import FrangTestCase, H2Config
+from test_suite.parameterize import param, parameterize
 
 
 class FrangLengthTestCase(FrangTestCase):
