+++ t_frang/test_whitelist_mark.py
@@ -1,9 +1,9 @@
 """Tests for Frang directive `whitelist_mark`."""
 
 from framework import curl_client, deproxy_client
-from framework.mixins import NetfilterMarkMixin
 from helpers import remote, tf_cfg
 from test_suite import tester
+from test_suite.mixins import NetfilterMarkMixin
 
 __author__ = "Tempesta Technologies, Inc."
 __copyright__ = "Copyright (C) 2024 Tempesta Technologies, Inc."
