+++ tls/test_tls_stress.py
@@ -2,8 +2,8 @@
 TLS Stress tests - load Tempesta FW with multiple TLS connections.
 """
 
-from framework.x509 import CertGenerator
 from helpers import dmesg, remote
+from helpers.x509 import CertGenerator
 from run_config import CONCURRENT_CONNECTIONS, DURATION, THREADS
 from test_suite import tester
